requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
dnspython==2.1.0
Flask==2.0.2
Flask-Cors==3.0.10
Flask-PyMongo==2.3.0
idna==3.3
imageio==2.13.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
networkx==2.6.3
numpy==1.21.4
packaging==21.3
pandas==1.3.4
Pillow==8.4.0
pymongo==3.12.1
pyparsing==3.0.6
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
PyWavelets==1.2.0
requests==2.26.0
scikit-image==0.19.0
scipy==1.7.3
six==1.16.0
tifffile==2021.11.2
torch==1.10.0
torchaudio==0.10.0
torchvision==0.11.1
tqdm==4.62.3
typing_extensions==4.0.1
urllib3==1.26.7
Werkzeug==2.0.2
